# src/docs/swagger.doc.json
{ 
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Match",
    "description": "Workout Match is an API for creating and managing customized exercises and workouts. With this API, you can create exercises with sets, reps, and exercise classification (type), as well as group them into workouts that can have a defined difficulty level. Additionally, Workout Match also offers the ability to search for exercises and workouts from other users and copy them to your own account, which is especially useful for those looking for new workout ideas and variations.",
    "contact": {
      "email": "arthurgbsf@gmail.com"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Database"
    },
    {
      "url": "https://workout-match.onrender.com/api-docs",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "parameters": [{
          "in":"query", 
          "name": "user", 
          "required": false,
          "schema":{"type": "boolean"},
        "description": "Optional. Pass 'user = true' to get the logged user information."}
        ],
        "description": "Route to get users information.",
        "tags": [
        "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "examples": {
                  "users": {
                    "value": {
                      "_id": "6447d9c97199225fa3f64dbf",
                      "name": "João da Silva",
                      "email": "joao@example.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Jwt Token expired.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "jwt expired"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Denied Acess. Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "Denied acess."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/users/user/{id}": {
      "get": {
        "parameters": [{
          "in":"path", 
          "name": "id", 
          "required": true,
          "schema":{"type": "string"},
        "description": "Pass an 'id' to get a thirdparty user information."}
        ],
        "description": "Route to get user iformation by 'id'.",
        "tags": [
        "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "examples": {
                  "users": {
                    "value": {
                      "_id": "6447d9c97199225fa3f64dbf",
                      "name": "João da Silva",
                      "email": "joao@example.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Jwt Token expired.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "jwt expired"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Denied Acess. Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "Denied acess."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/users/user": {
      "post": {
        "description": "Route to create an user.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "user": {
                  "value": {
                    "name": "João da Silva",
                    "email": "joao@example.com",
                    "password": "123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "User created."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Name, email and password are required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "Name, email and password are required."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Email have already registered.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "menssage": {
                    "value": {
                      "menssage": "Email have already registered."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Route to update the registered user.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "Update user": {
                  "value": {
                    "name": "João da Silva Barbosa",
                    "email": "joaobarbosa@example.com",
                    "password": "123456"
                  }
                },
                "Update user email": {
                  "value": {
                    "email": "joaobarbosa@example.com"
                  }
                },
                "Update user password": {
                  "value": {
                    "password": "123456"
                  }
                },
                "Update user name": {
                  "value": {
                    "name": "João da Silva Barbosa"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "User updated."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Is just allowed to update name email and password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "Is just allowed to update name email and password."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Email already registered.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                },
                "examples": {
                  "message 01": {
                    "value": {
                      "message": "Email already registered."
                    }
                  },
                  "message 02": {
                    "value": {
                      "message": "User not found."
                    }
                  },
                  "message 03": {
                    "value": {
                      "message": "User not modified."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Route to delete registered user.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The user was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "The user was deleted."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User wasn't deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "User wasn't deleted."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}